PDF encryption: https://github.com/metanorma/metanorma/issues/224
